+++ package.json
@@ -25,6 +25,7 @@
     "prop-types": "^15.6.0",
     "react": "^16.2.0",
     "react-dom": "^16.4.1",
+    "react-google-recaptcha": "^1.0.4",
     "react-helmet": "^5.2.0",
     "uuid": "^3.2.1"
   },
